debian: parse /etc/adduser.conf file manually
The syntax for /etc/adduser.conf allows spaces around the equals sign,
i.e.:

	GROUPHOMES = no

We were previously asking /bin/sh to source the file, which doesn't
support the above syntax.  So pull out the necessary fields via
sed/grep.

Thanks to Juan A. Diaz for reporting the bug and suggesting the fix.

Addresses-Debian-Bug: #487443

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>