resize2fs: Fix double bumping of directory in-use counts
When moving directories into new block groups (which would only happen
when shrinking a filesystem), resize2fs would increase the directory
in-use count by 2 times the necessary value, due to a change in
ext2fs_inode_alloc_stats() made in e2fsprogs 1.26.  This is largely
harmless, but it does result in a filesystem corruption for e2fsck to
fix.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>